pyproject.toml
[project]
name = "document-structuring-agent"
version = "0.1.0"
description = "Agentic document structuring pipeline: OCR HTML → structured document tree"
readme = "README.md"
authors = [
    { name = "Thomas Decloedt", email = "thomas.r.decloedt@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=1.59.0",
    "anthropic>=0.78.0,<0.80.0",  # >=0.78 needed for Claude 4.6; <0.80 avoids UserLocation import break in pydantic-ai (PR #4345 pending)
    "langfuse",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
